+++ examples/basic/main.tf
@@ -15,7 +15,7 @@ locals {
       master_ipv4_cidr_block       = "10.0.0.0/28"                              # CIDR block for the cluster control plane
       deletion_protection          = true                                       # (Optional) default is true
       enable_autopilot             = true
-      enable_backup                = true                                       # (Recommended)
+      enable_backup                = true # (Recommended)
       daily_maintenance_start_time = "00:00"
       # cluster_service_account_email = "" # (Recommended) this value should be filled out before cluster creation
     }
@@ -37,7 +37,7 @@ provider "google" {
 }
 
 module "gke-surrealdb" {
-  source = "dvanmali/surrealdb/google"
+  source  = "dvanmali/surrealdb/google"
   version = "1.2.3"
 
   project_id    = local.project_id